// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface CookieBannerDocumentDataCookieTypesItem {
  /**
   * Title field in *Cookie Banner → Cookie Types*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: cookie_banner.cookie_types[].title
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  title: prismic.KeyTextField;

  /**
   * Cookie ID field in *Cookie Banner → Cookie Types*
   *
   * - **Field Type**: Text
   * - **Placeholder**: For example "ad_storage"
   * - **API ID Path**: cookie_banner.cookie_types[].cookie_id
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  cookie_id: prismic.KeyTextField;

  /**
   * Description field in *Cookie Banner → Cookie Types*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: cookie_banner.cookie_types[].description
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  description: prismic.RichTextField;

  /**
   * Checked & Disabled field in *Cookie Banner → Cookie Types*
   *
   * - **Field Type**: Boolean
   * - **Placeholder**: *None*
   * - **Default Value**: false
   * - **API ID Path**: cookie_banner.cookie_types[].checked_and_disabled
   * - **Documentation**: https://prismic.io/docs/fields/boolean
   */
  checked_and_disabled: prismic.BooleanField;
}

/**
 * Content for Cookie Banner documents
 */
interface CookieBannerDocumentData {
  /**
   * Title field in *Cookie Banner*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: cookie_banner.title
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  title: prismic.RichTextField;

  /**
   * Description field in *Cookie Banner*
   *
   * - **Field Type**: Rich Text
   * - **Placeholder**: *None*
   * - **API ID Path**: cookie_banner.description
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/rich-text
   */
  description: prismic.RichTextField;

  /**
   * Accept Necessary Label field in *Cookie Banner*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: cookie_banner.accept_necessary_label
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  accept_necessary_label: prismic.KeyTextField;

  /**
   * Accept All Label field in *Cookie Banner*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: cookie_banner.accept_all_label
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  accept_all_label: prismic.KeyTextField;

  /**
   * Customize Label field in *Cookie Banner*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: cookie_banner.customize_label
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  customize_label: prismic.KeyTextField;

  /**
   * Cookie Types field in *Cookie Banner*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: cookie_banner.cookie_types[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/repeatable-group
   */
  cookie_types: prismic.GroupField<
    Simplify<CookieBannerDocumentDataCookieTypesItem>
  >;

  /**
   * Cancel Label field in *Cookie Banner*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: cookie_banner.cancel_label
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  cancel_label: prismic.KeyTextField;

  /**
   * Save Preferences Label field in *Cookie Banner*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: cookie_banner.save_preferences_label
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/fields/text
   */
  save_preferences_label: prismic.KeyTextField;
}

/**
 * Cookie Banner document from Prismic
 *
 * - **API ID**: `cookie_banner`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/content-modeling
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type CookieBannerDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithoutUID<
    Simplify<CookieBannerDocumentData>,
    'cookie_banner',
    Lang
  >;